Reject goal if action server is not ready
The previous implementation was falling into timeout while trying to
send a goal to an action server that is not present.

This updated implementation checks if the action server is ready to be
used. In case that the server is not ready, it tries to create a new
connection, if it fails to do so, it rejects setting the goal and
returns FAILURE.